Fix moveTo state storage bug
When a player selects a choice that depends on which album was chosen -
aka, an object, not a simple number - the whole object gets saved in the
"already played" storage. This doesn't work to re-compose the game
state: we need the numerical indices of all paths *chosen*.

The reason this happens is that the storage happens outside the check
for number/object.

To fix, re-factor this area to always save the real number in storage,
and simplify the logic.