Re-factor restarting to add a special message
When you restart the game, show a special message and add a horizontal
rule to make it more obvious you've restarted. Reset points and money.

Re-factor player get and set methods, so there is no falsey checks. It's
a nice idea, but sometimes you really want to override the value with a
falsey one. A guard at the implementation site is okay.